+++ examples/realworld-app/rw-database/src/main/java/realworld/entity/comment/CommentForListDTO.java
@@ -25,17 +25,17 @@ import realworld.entity.user.Author;
 import java.time.Instant;
 
 public record CommentForListDTO(
-        Integer id,
-        @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS", timezone = "UTC")
-        Instant createdAt,
-        @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS", timezone = "UTC")
-        Instant updatedAt,
-        String body,
-        Author author) {
+    Long id,
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS", timezone = "UTC")
+    Instant createdAt,
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS", timezone = "UTC")
+    Instant updatedAt,
+    String body,
+    Author author) {
 
     public CommentForListDTO(Comment comment) {
         this(comment.id(), comment.createdAt(),
-                comment.updatedAt(), comment.body(),
-                comment.author());
+            comment.updatedAt(), comment.body(),
+            comment.author());
     }
 }
